move instrument construction to apollo/instruments
